simplify FAB service interfaces and improve naming
