+++ tests/pre_award/fund_store_tests/test_fund_config_service.py
@@ -6,7 +6,7 @@ from unittest.mock import patch
 
 import pytest
 
-from pre_award.fund_store.services.save_fund_configuration import process_fund_config
+from pre_award.fund_store.services.fab_import_service import process_fund_config
 
 
 @pytest.fixture
@@ -30,13 +30,13 @@ def sample_converted_config():
 def test_process_fund_config_success(sample_converted_config):
     """Test successful fund config processing"""
     with (
-        patch("pre_award.fund_store.services.save_fund_configuration.insert_fund_data") as mock_insert_fund,
-        patch("pre_award.fund_store.services.save_fund_configuration.upsert_round_data") as mock_upsert_round,
-        patch("pre_award.fund_store.services.save_fund_configuration.insert_base_sections") as mock_insert_base,
+        patch("pre_award.fund_store.services.fab_import_service.insert_fund_data") as mock_insert_fund,
+        patch("pre_award.fund_store.services.fab_import_service.upsert_round_data") as mock_upsert_round,
+        patch("pre_award.fund_store.services.fab_import_service.insert_base_sections") as mock_insert_base,
         patch(
-            "pre_award.fund_store.services.save_fund_configuration.insert_or_update_application_sections"
+            "pre_award.fund_store.services.fab_import_service.insert_or_update_application_sections"
         ) as mock_insert_app,
-        patch("pre_award.fund_store.services.save_fund_configuration.db.session.commit") as mock_commit,
+        patch("pre_award.fund_store.services.fab_import_service.db.session.commit") as mock_commit,
     ):
         result = process_fund_config(sample_converted_config)
 
@@ -67,10 +67,10 @@ def test_process_fund_config_handles_errors(app, sample_converted_config):
     with app.app_context():
         with (
             patch(
-                "pre_award.fund_store.services.save_fund_configuration.insert_fund_data",
+                "pre_award.fund_store.services.fab_import_service.insert_fund_data",
                 side_effect=Exception("Database connection failed"),
             ) as mock_insert,
-            patch("pre_award.fund_store.services.save_fund_configuration.db.session.rollback") as mock_rollback,
+            patch("pre_award.fund_store.services.fab_import_service.db.session.rollback") as mock_rollback,
         ):
             result = process_fund_config(sample_converted_config)
